requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-widget-tweaks>=1.4.8
markdown>=3.3.4
python-decouple>=3.4
dj-database-url>=0.5.0
djangorestframework>=3.12.4